Update patch set 2
Patch Set 2:

I'm not seeing any kernel panics on my jbbl xt897, but I haven't enabled zram on it yet, so it may be related.

Could you please post a link to one of those panics' logs?

I'm not exactly against this change but rather undecided at this point...

Patch-set: 2