Update patch set 1
Patch Set 1:

Well, the stock firmware goes with 140 for such case.

I'd suggest to use a simple look-up table get the best keyboard back-light behaviour in relation to the display back-light.

Please propose an 256 member array mapping the display to the keyboard back-light that suits your needs. I'm sure we can agree on something that will serve well the majority of users in the end.

Patch-set: 1